refactor(renderer): settings store 拆出 catalog store（目录搜索/装卸/已配置包/面板 tab），依赖单向 catalog→settings
